requirements.txt
# ── ADK + Vertex AI ───────────────────────────────────────────────────────────
google-adk==1.2.1
google-cloud-aiplatform>=1.60.0
google-genai>=1.0.0

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0

# ── Config + validation ───────────────────────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Lint / dev tooling ────────────────────────────────────────────────────────
ruff>=0.6.0
